made update to endpoint for add to org button

+++ src/sentry/api/endpoints/organization_member/index.py
@@ -22,6 +22,7 @@ from sentry.apidocs.examples.organization_member_examples import OrganizationMem
 from sentry.apidocs.parameters import GlobalParams
 from sentry.apidocs.utils import inline_sentry_response_serializer
 from sentry.auth.authenticators import available_authenticators
+from sentry.auth.superuser import superuser_has_permission
 from sentry.models.integrations.external_actor import ExternalActor
 from sentry.models.organizationmember import InviteStatus, OrganizationMember
 from sentry.models.team import Team, TeamStatus
@@ -74,6 +75,9 @@ class MemberConflictValidationError(serializers.ValidationError):
 class OrganizationMemberRequestSerializer(serializers.Serializer):
     email = AllowedEmailField(
         max_length=75, required=True, help_text="The email address to send the invitation to."
+    )
+    user_id = serializers.CharField(
+        required=False, help_text="The user id to add to the organization. For superuser only."
     )
     role = serializers.ChoiceField(
         choices=roles.get_choices(), default=organization_roles.get_default().id
@@ -371,27 +375,39 @@ class OrganizationMemberIndexEndpoint(OrganizationEndpoint):
                 status=400,
             )
 
-        with transaction.atomic(router.db_for_write(OrganizationMember)):
-            # remove any invitation requests for this email before inviting
-            existing_invite = OrganizationMember.objects.filter(
-                Q(invite_status=InviteStatus.REQUESTED_TO_BE_INVITED.value)
-                | Q(invite_status=InviteStatus.REQUESTED_TO_JOIN.value),
-                email=result["email"],
-                organization=organization,
-            )
-            for om in existing_invite:
-                om.delete()
+        if superuser_has_permission(request):
+            with transaction.atomic(router.db_for_write(OrganizationMember)):
+                om = OrganizationMember.objects.create(
+                    organization=organization,
+                    user_email=result["email"],
+                    user_id=result["user_id"],
+                    role=result["role"],
+                    inviter_id=request.user.id,
+                    invite_status=InviteStatus.APPROVED.value,
+                )
+                om.save()
+        else:
+            with transaction.atomic(router.db_for_write(OrganizationMember)):
+                # remove any invitation requests for this email before inviting
+                existing_invite = OrganizationMember.objects.filter(
+                    Q(invite_status=InviteStatus.REQUESTED_TO_BE_INVITED.value)
+                    | Q(invite_status=InviteStatus.REQUESTED_TO_JOIN.value),
+                    email=result["email"],
+                    organization=organization,
+                )
+                for om in existing_invite:
+                    om.delete()
 
-            om = OrganizationMember(
-                organization=organization,
-                email=result["email"],
-                role=result["role"],
-                inviter_id=request.user.id,
-            )
+                om = OrganizationMember(
+                    organization=organization,
+                    email=result["email"],
+                    role=result["role"],
+                    inviter_id=request.user.id,
+                )
 
-            if settings.SENTRY_ENABLE_INVITES:
-                om.token = om.generate_token()
-            om.save()
+                if settings.SENTRY_ENABLE_INVITES:
+                    om.token = om.generate_token()
+                om.save()
 
         # Do not set team-roles when inviting members
         if "teamRoles" in result or "teams" in result:
@@ -402,23 +418,32 @@ class OrganizationMemberIndexEndpoint(OrganizationEndpoint):
             )
             save_team_assignments(om, teams)
 
-        if settings.SENTRY_ENABLE_INVITES and result.get("sendInvite"):
-            referrer = request.query_params.get("referrer")
-            om.send_invite_email(referrer)
-            member_invited.send_robust(
-                member=om, user=request.user, sender=self, referrer=request.data.get("referrer")
+        if superuser_has_permission(request):
+            self.create_audit_entry(
+                request=request,
+                organization_id=organization.id,
+                target_object=om.id,
+                data=om.get_audit_log_data(),
+                event=(audit_log.get_event_id("MEMBER_ADD")),
             )
 
-        self.create_audit_entry(
-            request=request,
-            organization_id=organization.id,
-            target_object=om.id,
-            data=om.get_audit_log_data(),
-            event=(
-                audit_log.get_event_id("MEMBER_INVITE")
-                if settings.SENTRY_ENABLE_INVITES
-                else audit_log.get_event_id("MEMBER_ADD")
-            ),
-        )
+        else:
+            if settings.SENTRY_ENABLE_INVITES and result.get("sendInvite"):
+                referrer = request.query_params.get("referrer")
+                om.send_invite_email(referrer)
+                member_invited.send_robust(
+                    member=om, user=request.user, sender=self, referrer=request.data.get("referrer")
+                )
+            self.create_audit_entry(
+                request=request,
+                organization_id=organization.id,
+                target_object=om.id,
+                data=om.get_audit_log_data(),
+                event=(
+                    audit_log.get_event_id("MEMBER_INVITE")
+                    if settings.SENTRY_ENABLE_INVITES
+                    else audit_log.get_event_id("MEMBER_ADD")
+                ),
+            )
 
         return Response(serialize(om), status=201)
